fix(engine): 64-bit offset_dst in all 11 vendored MMQ launchers (audit Q7)
The tile-output offset 'jt * mmq_x * stride_col_dst + it * mmq_y' computed
entirely in int32 and fed a pointer add — wrapping negative once one MMQ call
covers n_tokens * out_f >= 2^31 (T >= 8193 at a 256k vocab: the quantized
lm_head through the full-T prompt-logits path; jt = 64 at mmq_x = 128 is
exactly 2^31). The FlashInfer #4263 remediation: widen the product to int64_t
at every launcher.

Sites (11): mmq_q8_0.cu:387, mmq_q8_0_f32acc.cu:414, mmq_q4_0.cu:445/514/581/604
(static + CLC + stream-k interior + stream-k fixup), mmq_q45k.cu:481,
mmq_nvfp4_w4a8.cu:788/1367, mmq_fp4.cu:435, mmq_fp8_blk.cu:422.

Also widened (defense-in-depth, audit-flagged latent): quantize_q8_1 /
quantize_fp4_act global thread-id math in qmatvec.cu — int32-safe today
(needs M >= ~52k tokens), free to widen while touching the area.

No behavior change at supported shapes: the multiply widens, small-shape
results are bit-identical (kernel-check gate below). Serve hot paths never
reach the bound (prime chunks at 4096, decode lm_head is last-row-only);
the exposed entry is the dense prompt-logits eval path.

Audit: research/sweep-audits-20260805/AUDIT.md §4 (brief Q7).
Gate: kernel-check ALL GREEN on the 5090 (receipt in
research/audit-fixes-20260805/).

+++ crates/memra-engine/cu/mmq_fp8_blk.cu
@@ -419,7 +419,8 @@ static __global__ void mul_mat_q_fp8_blk(
     const int it = blockIdx.x;   // out-row tile; scale grid ROW == (it*mmq_y)>>7 (mmq_y | FP8_BLK)
 
     const int offset_y     = (jt * mmq_x) * (sizeof(block_e4m3_mmq) / sizeof(int));
-    const int offset_dst   = jt * mmq_x * stride_col_dst + it * mmq_y;
+    // 64-bit offset_dst (audit Q7, 2026-08-05): wraps at n_tokens*out_f >= 2^31 — see mmq_q8_0.cu.
+    const int64_t offset_dst = (int64_t) jt * mmq_x * stride_col_dst + (int64_t) it * mmq_y;
     const int tile_x_max_i = nrows_x   - it * mmq_y - 1;
     const int tile_y_max_j = ncols_dst - jt * mmq_x - 1;
